fix: added a new component and optimized props transfer.
